add mobile and desktop versions of car scroller
